Añadir gestión de movimientos de productos
Se agregó la vista `ListaMovimientosView` para gestionar movimientos
de productos (entradas, salidas, etc.), con funcionalidades para
listar, agregar, modificar, eliminar y buscar movimientos.

Cambios principales:
- Se creó el servicio `MovimientoService` para operaciones CRUD
  con la API REST.
- Se actualizaron los modelos `Movimiento` y `Producto` para
  incluir nuevas propiedades (`id_producto`, `cantidad`, `fecha`).
- Se corrigió el tipo de `presio` a `precio` en `Producto`.
- Se añadieron controles y lógica en `ListProductosView` para
  manejar movimientos.
- Se diseñó la interfaz gráfica y lógica de `ListaMovimientosView`.

Estos cambios introducen una nueva funcionalidad para gestionar
movimientos y actualizar el stock de productos.

+++ ControlDeStock/Models/Producto.cs
@@ -11,7 +11,7 @@ namespace ControlDeStock.Models
         public string? _id { get; set; }
         public string nombre { get; set; }
         public string categoria { get; set; }
-        public int presio { get; set; }
+        public decimal presio { get; set; }
         public int stock_actual { get; set; }
     }
 }
